@@ -4,7 +4,7 @@ description: A new Peercoin wallet.
 version: 1.1.3+124
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 publish_to: none
 
@@ -55,6 +55,7 @@ dependencies:
   grpc: ^3.1.0
   mockito: ^5.3.2
   cryptography_flutter: ^2.0.2
+  loader_overlay: ^2.1.0
 
 dev_dependencies:
   integration_test:
