@@ -3,36 +3,17 @@ import 'package:flutter/material.dart';
 class MyTheme {
   static Map<ThemeMode, ThemeData> appThemes = {
     ThemeMode.light: ThemeData(
-      colorScheme: ColorScheme(
-        primary: LightColors.green,
-        primaryContainer: LightColors.darkGreen,
-        secondary: LightColors.grey,
-        secondaryContainer: LightColors.lightGrey,
-        surface: LightColors.white,
-        background: LightColors.grey,
-        error: LightColors.red,
-        onPrimary: LightColors.white,
-        onSecondary: LightColors.blackGreen,
-        onSurface: LightColors.green,
-        onBackground: LightColors.green,
-        onError: LightColors.red,
-        brightness: Brightness.light,
-      ),
-      backgroundColor: LightColors.white,
-      bottomAppBarColor: const Color(0x003cb054),
       cardColor: LightColors.white,
       dialogBackgroundColor: LightColors.white,
       disabledColor: LightColors.lightGreen,
       dividerColor: LightColors.darkGreen,
-      errorColor: LightColors.red,
       focusColor: LightColors.green,
       hintColor: LightColors.grey,
       primaryColor: LightColors.green,
-      primarySwatch: materialColor(LightColors.grey),
       shadowColor: LightColors.lightGreen,
       unselectedWidgetColor: LightColors.grey,
       textTheme: TextTheme(
-        button: TextStyle(
+        labelLarge: TextStyle(
           letterSpacing: 1.4,
           fontSize: 16,
           color: DarkColors.white,
@@ -68,30 +49,30 @@ class MyTheme {
           foregroundColor: LightColors.green,
         ),
       ),
+      colorScheme: ColorScheme(
+        primary: LightColors.green,
+        primaryContainer: LightColors.darkGreen,
+        secondary: LightColors.grey,
+        secondaryContainer: LightColors.lightGrey,
+        surface: LightColors.white,
+        background: LightColors.grey,
+        error: LightColors.red,
+        onPrimary: LightColors.white,
+        onSecondary: LightColors.blackGreen,
+        onSurface: LightColors.green,
+        onBackground: LightColors.green,
+        onError: LightColors.red,
+        brightness: Brightness.light,
+      )
+          .copyWith(background: LightColors.white)
+          .copyWith(error: LightColors.red),
+      bottomAppBarTheme: const BottomAppBarTheme(color: Color(0x003cb054)),
     ),
     ThemeMode.dark: ThemeData.dark().copyWith(
-      colorScheme: ColorScheme(
-        primary: DarkColors.green,
-        primaryContainer: DarkColors.white,
-        secondary: DarkColors.white,
-        secondaryContainer: DarkColors.grey,
-        surface: DarkColors.black,
-        background: DarkColors.green,
-        error: DarkColors.red,
-        onPrimary: DarkColors.green,
-        onSecondary: DarkColors.green,
-        onSurface: DarkColors.green,
-        onBackground: DarkColors.green,
-        onError: DarkColors.red,
-        brightness: Brightness.dark,
-      ),
-      backgroundColor: DarkColors.green,
-      bottomAppBarColor: const Color(0x000D1821),
       cardColor: DarkColors.green,
       dialogBackgroundColor: DarkColors.green,
       disabledColor: DarkColors.darkBlue,
       dividerColor: DarkColors.lightGreen,
-      errorColor: DarkColors.red,
       focusColor: DarkColors.black,
       hintColor: DarkColors.white,
       primaryColor: DarkColors.black,
@@ -105,17 +86,17 @@ class MyTheme {
       sliderTheme: SliderThemeData(valueIndicatorColor: DarkColors.darkBlue),
       unselectedWidgetColor: DarkColors.white,
       textTheme: TextTheme(
-        headline6: TextStyle(color: DarkColors.white),
-        headline5: TextStyle(color: DarkColors.white),
-        headline4: TextStyle(color: DarkColors.white),
-        headline3: TextStyle(color: DarkColors.white),
-        headline2: TextStyle(color: DarkColors.white),
-        headline1: TextStyle(color: DarkColors.white),
-        subtitle1: TextStyle(color: DarkColors.white),
-        subtitle2: TextStyle(color: DarkColors.white),
-        bodyText1: TextStyle(color: DarkColors.white),
-        bodyText2: TextStyle(color: DarkColors.white),
-        button: TextStyle(
+        titleLarge: TextStyle(color: DarkColors.white),
+        headlineSmall: TextStyle(color: DarkColors.white),
+        headlineMedium: TextStyle(color: DarkColors.white),
+        displaySmall: TextStyle(color: DarkColors.white),
+        displayMedium: TextStyle(color: DarkColors.white),
+        displayLarge: TextStyle(color: DarkColors.white),
+        titleMedium: TextStyle(color: DarkColors.white),
+        titleSmall: TextStyle(color: DarkColors.white),
+        bodyLarge: TextStyle(color: DarkColors.white),
+        bodyMedium: TextStyle(color: DarkColors.white),
+        labelLarge: TextStyle(
           letterSpacing: 1.4,
           fontSize: 16,
           color: DarkColors.white,
@@ -156,6 +137,22 @@ class MyTheme {
       textSelectionTheme: TextSelectionThemeData(
         cursorColor: DarkColors.white,
       ),
+      colorScheme: ColorScheme(
+        primary: DarkColors.green,
+        primaryContainer: DarkColors.white,
+        secondary: DarkColors.white,
+        secondaryContainer: DarkColors.grey,
+        surface: DarkColors.black,
+        background: DarkColors.green,
+        error: DarkColors.red,
+        onPrimary: DarkColors.green,
+        onSecondary: DarkColors.green,
+        onSurface: DarkColors.green,
+        onBackground: DarkColors.green,
+        onError: DarkColors.red,
+        brightness: Brightness.dark,
+      ).copyWith(background: DarkColors.green).copyWith(error: DarkColors.red),
+      bottomAppBarTheme: const BottomAppBarTheme(color: Color(0x000D1821)),
     )
   };
 
