@@ -2,6 +2,7 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:loader_overlay/loader_overlay.dart';
 import 'package:provider/provider.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
@@ -81,6 +82,7 @@ class _WalletHomeState extends State<WalletHomeScreen>
 
   @override
   void dispose() {
+    _activeWallets.closeWallet(_wallet.name);
     WidgetsBinding.instance.removeObserver(this);
     super.dispose();
   }
@@ -117,10 +119,13 @@ class _WalletHomeState extends State<WalletHomeScreen>
       await _activeWallets.generateUnusedAddress(_wallet.name);
       _walletTransactions =
           await _activeWallets.getWalletTransactions(_wallet.name);
-      await _connectionProvider!
-          .init(_wallet.name, requestedFromWalletHome: true);
+      await _connectionProvider!.init(
+        _wallet.name,
+        requestedFromWalletHome: true,
+      );
 
       if (_appSettings.authenticationOptions!['walletHome']!) {
+        // ignore: use_build_context_synchronously
         await Auth.requireAuth(
           context: context,
           biometricsAllowed: _appSettings.biometricsAllowed,
@@ -137,6 +142,8 @@ class _WalletHomeState extends State<WalletHomeScreen>
       }
 
       checkPendingNotifications();
+      // ignore: use_build_context_synchronously
+      context.loaderOverlay.hide();
 
       if (arguments.containsKey('pushedAddress')) {
         changeIndex(Tabs.send, arguments['pushedAddress']);
@@ -228,6 +235,7 @@ class _WalletHomeState extends State<WalletHomeScreen>
               ) >=
               1000) {
         //Coins worth 1000 USD or more
+        // ignore: use_build_context_synchronously
         await showDialog(
           context: context,
           builder: (_) => AlertDialog(
