@@ -23,7 +23,7 @@ class SendTabAddressManagement extends StatelessWidget {
                 onPressed: () => onDelete(),
                 icon: Icon(
                   Icons.delete,
-                  color: Theme.of(context).errorColor,
+                  color: Theme.of(context).colorScheme.error,
                 ),
               )
             : const SizedBox(
